+++ exchange/src/exchange/dashboard/state/ticker_state.hpp
@@ -17,33 +17,33 @@ struct BotStates {
     size_t min_open_asks_{};
     size_t max_open_bids_{};
     size_t max_open_asks_{};
-    float min_utilization_{};
-    float max_utilization_{};
-    float min_pnl_{};
-    float max_pnl_{};
+    double min_utilization_{};
+    double max_utilization_{};
+    double min_pnl_{};
+    double max_pnl_{};
 
-    size_t avg_open_bids_{};
-    size_t avg_open_asks_{};
-    float avg_bid_interest_{};
-    float avg_ask_interest_{};
-    float avg_utilization_{};
-    float avg_pnl_{};
+    double avg_open_bids_{};
+    double avg_open_asks_{};
+    double avg_bid_interest_{};
+    double avg_ask_interest_{};
+    double avg_utilization_{};
+    double avg_pnl_{};
 };
 
 struct TickerState : public ticks::TickObserver {
     const std::string TICKER;
-    const float STARTING_PRICE;
+    const double STARTING_PRICE;
 
-    float theo_{};
-    float midprice_{};
-    std::pair<float, float> spread_{};
+    double theo_{};
+    double midprice_{};
+    std::pair<double, double> spread_{};
     size_t num_bids_{};
     size_t num_asks_{};
 
     BotStates mm_state_{};
     BotStates retail_state_{};
 
-    TickerState(std::string ticker, float starting_price) :
+    TickerState(std::string ticker, double starting_price) :
         TICKER(std::move(ticker)), STARTING_PRICE(starting_price)
     {}
 
@@ -65,10 +65,10 @@ private:
             bot_states.min_open_asks_ = INFINITY;
             bot_states.max_open_bids_ = -INFINITY;
             bot_states.max_open_asks_ = -INFINITY;
-            bot_states.min_utilization_ = static_cast<float>(INFINITY);
-            bot_states.max_utilization_ = -static_cast<float>(INFINITY);
-            bot_states.min_pnl_ = static_cast<float>(INFINITY);
-            bot_states.max_pnl_ = -static_cast<float>(INFINITY);
+            bot_states.min_utilization_ = static_cast<double>(INFINITY);
+            bot_states.max_utilization_ = -static_cast<double>(INFINITY);
+            bot_states.min_pnl_ = static_cast<double>(INFINITY);
+            bot_states.max_pnl_ = -static_cast<double>(INFINITY);
             bot_states.avg_open_bids_ = 0;
             bot_states.avg_open_asks_ = 0;
             bot_states.avg_bid_interest_ = 0;
