+++ exchange/src/exchange/traders/trader_manager.cpp
@@ -6,8 +6,8 @@ namespace manager {
 std::optional<messages::SIDE>
 ClientManager::validate_match(const messages::Match& match) const
 {
-    float trade_value = match.price * match.quantity;
-    float remaining_capital = get_capital(match.buyer_id) - trade_value;
+    double trade_value = match.price * match.quantity;
+    double remaining_capital = get_capital(match.buyer_id) - trade_value;
 
     bool insufficient_holdings =
         get_holdings(match.seller_id, match.ticker) - match.quantity < 0;
